conversational-bi-dashboard/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pandas==2.1.3
python-multipart==0.0.6
google-generativeai==0.3.0
python-dotenv==1.0.0
sqlalchemy==2.0.23
